1.腳本重啟會自動刪爆炸前帳號 2.Main增加多語言支持，優化player not found 3.增加Main2，位置為右上角 4.SCALE100支持(可能有問題) 5.出神包更換頭像成皮卡丘以及簽名改成我是新人，建號時頭像設置成坂木 6.settings出廠設置，webhook等默認設置完成，默認刪除賬號模式 7.合併了6.3.9 8.DeadHBWebhookURL可以在啟動界面配置

+++ Scripts/1.ahk
@@ -15,7 +15,7 @@ CoordMode, Pixel, Screen
 DllCall("AllocConsole")
 WinHide % "ahk_id " DllCall("GetConsoleWindow", "ptr")
 
-global winTitle, changeDate, failSafe, openPack, Delay, failSafeTime, StartSkipTime, Columns, failSafe, adbPort, scriptName, adbShell, adbPath, GPTest, StatusText, defaultLanguage, setSpeed, jsonFileName, pauseToggle, SelectedMonitorIndex, swipeSpeed, godPack, scaleParam, discordUserId, discordWebhookURL, deleteMethod, packs, FriendID, friendIDs, Instances, username, friendCode, stopToggle, friended, runMain, showStatus, injectMethod, packMethod, loadDir, loadedAccount, nukeAccount, TrainerCheck, FullArtCheck, RainbowCheck, dateChange, foundGP, foundTS, friendsAdded, minStars, PseudoGodPack, Palkia, Dialga, Mew, Pikachu, Charizard, Mewtwo, packArray, CrownCheck, ImmersiveCheck, slowMotion, screenShot, accountFile, invalid, starCount, gpFound, foundTS
+global winTitle, changeDate, failSafe, openPack, Delay, failSafeTime, StartSkipTime, Columns, failSafe, adbPort, scriptName, adbShell, adbPath, GPTest, StatusText, defaultLanguage, setSpeed, jsonFileName, pauseToggle, SelectedMonitorIndex, swipeSpeed, godPack, scaleParam, discordUserId, discordWebhookURL, deleteMethod, packs, FriendID, friendIDs, Instances, username, friendCode, stopToggle, friended, runMain, showStatus, injectMethod, packMethod, loadDir, loadedAccount, nukeAccount, TrainerCheck, FullArtCheck, RainbowCheck, dateChange, foundGP, foundTS, friendsAdded, minStars, PseudoGodPack, Palkia, Dialga, Mew, Pikachu, Charizard, Mewtwo, packArray, CrownCheck, ImmersiveCheck, slowMotion, DeadCheck, heartBeatName, screenShot, accountFile, invalid, starCount, gpFound, foundTS
 scriptName := StrReplace(A_ScriptName, ".ahk")
 winTitle := scriptName
 foundGP := false
@@ -59,6 +59,7 @@ IniRead, Pikachu, %A_ScriptDir%\..\Settings.ini, UserSettings, Pikachu, 0
 IniRead, Charizard, %A_ScriptDir%\..\Settings.ini, UserSettings, Charizard, 0
 IniRead, Mewtwo, %A_ScriptDir%\..\Settings.ini, UserSettings, Mewtwo, 0
 IniRead, slowMotion, %A_ScriptDir%\..\Settings.ini, UserSettings, slowMotion, 0
+IniRead, DeadCheck, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck, 0
 
 packArray :=[]
 
@@ -183,111 +184,119 @@ if(!injectMethod || !loadedAccount)
 
 pToken := Gdip_Startup()
 packs := 0
+if(DeadCheck==1) {
+	friended:= true
+	menuDeleteStart()
+	IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
+	Reload
+}else{
+	Loop {
+		Randmax := packArray.Length()
+		Random, rand, 1, Randmax
+		openPack := packArray[rand]
+		friended := false
+		IniWrite, 1, %A_ScriptDir%\..\HeartBeat.ini, HeartBeat, Instance%scriptName%
+		FormatTime, CurrentTime,, HHmm
 
-Loop {
-	Randmax := packArray.Length()
-	Random, rand, 1, Randmax
-	openPack := packArray[rand]
-	friended := false
-	IniWrite, 1, %A_ScriptDir%\..\HeartBeat.ini, HeartBeat, Instance%scriptName%
-	FormatTime, CurrentTime,, HHmm
+		StartTime := changeDate - 45 ; 12:55 AM2355
+		EndTime := changeDate + 5 ; 1:01 AM
 
-	StartTime := changeDate - 45 ; 12:55 AM2355
-	EndTime := changeDate + 5 ; 1:01 AM
+		; Adjust for crossing midnight
+		if (StartTime < 0)
+			StartTime += 2400
+		if (EndTime >= 2400)
+			EndTime -= 2400
 
-	; Adjust for crossing midnight
-	if (StartTime < 0)
-		StartTime += 2400
-	if (EndTime >= 2400)
-		EndTime -= 2400
+		Random, randomTime, 3, 7
 
-	Random, randomTime, 3, 7
+		While(((CurrentTime - StartTime >= 0) && (CurrentTime - StartTime <= randomTime)) || ((EndTime - CurrentTime >= 0) && (EndTime - CurrentTime <= randomTime)))
+		{
+			CreateStatusMessage("I need a break... Sleeping until " . changeDate + randomTime . " `nto avoid being kicked out from the date change")
+			FormatTime, CurrentTime,, HHmm ; Update the current time after sleep
+			Sleep, 5000
+			dateChange := true
+		}
+		if(dateChange)
+			createAccountList(scriptName)
+		FindImageAndClick(65, 195, 100, 215, , "Platin", 18, 109, 2000) ; click mod settings
+		if(setSpeed = 3)
+			FindImageAndClick(182, 170, 194, 190, , "Three", 187, 180) ; click mod settings
+		else
+			FindImageAndClick(100, 170, 113, 190, , "Two", 107, 180) ; click mod settings
+		Delay(1)
+		adbClick(41, 296)
+		Delay(1)
+		packs := 0
 
-	While(((CurrentTime - StartTime >= 0) && (CurrentTime - StartTime <= randomTime)) || ((EndTime - CurrentTime >= 0) && (EndTime - CurrentTime <= randomTime)))
-	{
-		CreateStatusMessage("I need a break... Sleeping until " . changeDate + randomTime . " `nto avoid being kicked out from the date change")
-		FormatTime, CurrentTime,, HHmm ; Update the current time after sleep
-		Sleep, 5000
-		dateChange := true
-	}
-	if(dateChange)
-		createAccountList(scriptName)
-	FindImageAndClick(65, 195, 100, 215, , "Platin", 18, 109, 2000) ; click mod settings
-	if(setSpeed = 3)
-		FindImageAndClick(182, 170, 194, 190, , "Three", 187, 180) ; click mod settings
-	else
-		FindImageAndClick(100, 170, 113, 190, , "Two", 107, 180) ; click mod settings
-	Delay(1)
-	adbClick(41, 296)
-	Delay(1)
-	packs := 0
+		if(!injectMethod || !loadedAccount)
+			DoTutorial()
 
-	if(!injectMethod || !loadedAccount)
-		DoTutorial()
+		if(deleteMethod = "5 Pack" || packMethod)
+			if(!loadedAccount)
+				wonderPicked := DoWonderPick()
 
-	if(deleteMethod = "5 Pack" || packMethod)
-		if(!loadedAccount)
-			wonderPicked := DoWonderPick()
-
-	friendsAdded := AddFriends()
-	SelectPack("First")
-	PackOpening()
-
-	if(packMethod) {
-		friendsAdded := AddFriends(true)
-		SelectPack()
-	}
-
-	PackOpening()
-
-	if(!injectMethod || !loadedAccount)
-		HourglassOpening() ;deletemethod check in here at the start
-
-	if(wonderPicked) {
-		friendsAdded := AddFriends(true)
-		SelectPack("HGPack")
+		friendsAdded := AddFriends()
+		SelectPack("First")
 		PackOpening()
+
 		if(packMethod) {
+			friendsAdded := AddFriends(true)
+			SelectPack()
+		}
+
+		PackOpening()
+
+		if(!injectMethod || !loadedAccount)
+			HourglassOpening() ;deletemethod check in here at the start
+
+		if(wonderPicked) {
 			friendsAdded := AddFriends(true)
 			SelectPack("HGPack")
 			PackOpening()
+			if(packMethod) {
+				friendsAdded := AddFriends(true)
+				SelectPack("HGPack")
+				PackOpening()
+			}
+			else {
+				HourglassOpening(true)
+			}
 		}
-		else {
-			HourglassOpening(true)
+
+		if(nukeAccount && !injectMethod) {
+			menuDelete()
+			IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
+		}else{
+			RemoveFriends()
+			IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 		}
+		if(injectMethod)
+			loadedAccount := loadAccount()
+
+		if(!injectMethod || !loadedAccount) {
+			if(!nukeAccount) {
+				saveAccount("All")
+				restartGameInstance("New Run", false)
+			}
+		}
+
+		CreateStatusMessage("New Run")
+		rerolls++
+		if(!loadedAccount)
+			if(deleteMethod = "5 Pack" || packMethod)
+				packs := 5
+		AppendToJsonFile(packs)
+		totalSeconds := Round((A_TickCount - rerollTime) / 1000) ; Total time in seconds
+		avgtotalSeconds := Round(totalSeconds / rerolls) ; Total time in seconds
+		minutes := Floor(avgtotalSeconds / 60) ; Total minutes
+		seconds := Mod(avgtotalSeconds, 60) ; Remaining seconds within the minute
+		mminutes := Floor(totalSeconds / 60) ; Total minutes
+		sseconds := Mod(totalSeconds, 60) ; Remaining seconds within the minute
+		CreateStatusMessage("Avg: " . minutes . "m " . seconds . "s Runs: " . rerolls, 25, 0, 510)
+		LogToFile("Packs: " . packs . " Total time: " . mminutes . "m " . sseconds . "s Avg: " . minutes . "m " . seconds . "s Runs: " . rerolls)
+		if(stopToggle)
+			ExitApp
 	}
-
-	if(nukeAccount && !injectMethod)
-		menuDelete()
-	else
-		RemoveFriends()
-
-	if(injectMethod)
-		loadedAccount := loadAccount()
-
-	if(!injectMethod || !loadedAccount) {
-		if(!nukeAccount) {
-			saveAccount("All")
-			restartGameInstance("New Run", false)
-		}
-	}
-
-	CreateStatusMessage("New Run")
-	rerolls++
-	if(!loadedAccount)
-		if(deleteMethod = "5 Pack" || packMethod)
-			packs := 5
-	AppendToJsonFile(packs)
-	totalSeconds := Round((A_TickCount - rerollTime) / 1000) ; Total time in seconds
-	avgtotalSeconds := Round(totalSeconds / rerolls) ; Total time in seconds
-	minutes := Floor(avgtotalSeconds / 60) ; Total minutes
-	seconds := Mod(avgtotalSeconds, 60) ; Remaining seconds within the minute
-	mminutes := Floor(totalSeconds / 60) ; Total minutes
-	sseconds := Mod(totalSeconds, 60) ; Remaining seconds within the minute
-	CreateStatusMessage("Avg: " . minutes . "m " . seconds . "s Runs: " . rerolls, 25, 0, 510)
-	LogToFile("Packs: " . packs . " Total time: " . mminutes . "m " . sseconds . "s Avg: " . minutes . "m " . seconds . "s Runs: " . rerolls)
-	if(stopToggle)
-		ExitApp
 }
 return
 
@@ -425,7 +434,8 @@ TradeTutorial() {
 }
 
 AddFriends(renew := false, getFC := false) {
-	global FriendID, friendIds, waitTime, friendCode
+	global FriendID, friendIds, waitTime, friendCode, scriptName
+	IniWrite, 1, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 	friendIDs := ReadFile("ids")
 	count := 0
 	friended := true
@@ -595,6 +605,26 @@ AddFriends(renew := false, getFC := false) {
 	return n ;return added friends so we can dynamically update the .txt in the middle of a run without leaving friends at the end
 }
 
+changeProfile() {
+	FindImageAndClick(120, 500, 155, 530, , "Social", 143, 518, 500)
+	FindImageAndClick(20, 500, 55, 530, , "Home", 40, 516, 500) 212 276 230 294
+	FindImageAndClick(203, 272, 237, 300, , "Profile", 143, 95, 500)
+	FindImageAndClick(130, 448, 157, 457, , "OK7", 143, 179, 500)
+	Delay(3)
+	adbClick(143,244)
+	Delay(3)
+	adbClick(55,244)
+	Delay(3)
+	adbClick(143,466)
+	Delay(1)
+	FindImageAndClick(130, 448, 157, 457, , "OK7", 143, 316, 500)
+	Delay(3)
+	adbClick(143,244)
+	Delay(3)
+	adbClick(143,466)
+	Delay(1)
+}
+
 EraseInput(num := 0, total := 0) {
 	if(num)
 		CreateStatusMessage("Removing friend ID " . num . "/" . total)
@@ -679,6 +709,7 @@ FindOrLoseImage(X1, Y1, X2, Y2, searchVariation := "", imageName := "DEFAULT", E
 			CreateStatusMessage("Loaded deleted account. Deleting XML." )
 			if(loadedAccount) {
 				FileDelete, %loadedAccount%
+				IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 			}
 			LogToFile("Restarted game for instance " scriptName " Reason: No save data found", "Restart.txt")
 			Reload
@@ -823,6 +854,7 @@ FindImageAndClick(X1, Y1, X2, Y2, searchVariation := "", imageName := "DEFAULT",
 				CreateStatusMessage("Loaded deleted account. Deleting XML." )
 				if(loadedAccount) {
 					FileDelete, %loadedAccount%
+					IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 				}
 				LogToFile("Restarted game for instance " scriptName " Reason: No save data found", "Restart.txt")
 				Reload
@@ -946,15 +978,17 @@ waitadb() {
 }
 
 restartGameInstance(reason, RL := true){
-	global Delay, scriptName, adbShell, adbPath, adbPort, friended, loadedAccount
+	global Delay, scriptName, adbShell, adbPath, adbPort, friended, loadedAccount, DeadCheck
 	;initializeAdbShell()
 	CreateStatusMessage("Restarting game reason: `n" reason)
 
 	if(!RL || RL != "GodPack") {
 		adbShell.StdIn.WriteLine("am force-stop jp.pokemon.pokemontcgp")
 		waitadb()
-		if(!RL)
-			adbShell.StdIn.WriteLine("rm /data/data/jp.pokemon.pokemontcgp/shared_prefs/deviceAccount:.xml") ; delete account data
+		if(!RL){
+			if(DeadCheck==0)
+				adbShell.StdIn.WriteLine("rm /data/data/jp.pokemon.pokemontcgp/shared_prefs/deviceAccount:.xml") ; delete account data
+		}
 		;adbShell.StdIn.WriteLine("rm -rf /data/data/jp.pokemon.pokemontcgp/cache/*") ; clear cache
 		waitadb()
 		adbShell.StdIn.WriteLine("am start -n jp.pokemon.pokemontcgp/com.unity3d.player.UnityPlayerActivity")
@@ -964,9 +998,11 @@ restartGameInstance(reason, RL := true){
 
 	if(RL = "GodPack") {
 		LogToFile("Restarted game for instance " scriptName " Reason: " reason, "Restart.txt")
+		IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 		Reload
 	} else if(RL) {
 		if(menuDeleteStart()) {
+			IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 			logMessage := "\n" . username . "\n[" . starCount . "/5][" . packs . "P] " . invalid . " God pack found in instance: " . scriptName . "\nFile name: " . accountFile . "\nGot stuck getting friend code."
 			LogToFile(logMessage, "GPlog.txt")
 			LogToDiscord(logMessage, screenShot, discordUserId)
@@ -1143,6 +1179,7 @@ CreateStatusMessage(Message, GuiName := 50, X := 0, Y := 80) {
 }
 
 CheckPack() {
+	global scriptName, DeadCheck
 	foundGP := false ;check card border to find godpacks
 	foundTrainer := false
 	foundRainbow := false
@@ -1183,8 +1220,10 @@ CheckPack() {
 			foundTS := "Double two star"
 	}
 	if(foundGP || foundTrainer || foundRainbow || foundFullArt || foundImmersive || foundCrown || 2starCount > 1) {
-		if(loadedAccount)
+		if(loadedAccount){
 			FileDelete, %loadedAccount% ;delete xml file from folder if using inject method
+			IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
+		}
 		if(foundGP)
 			restartGameInstance("God Pack found. Continuing...", "GodPack") ; restarts to backup and delete xml file with account info.
 		else {
@@ -1195,9 +1234,11 @@ CheckPack() {
 }
 
 FoundStars(star) {
+	global scriptName, DeadCheck
 	screenShot := Screenshot(star)
 	accountFile := saveAccount(star)
 	friendCode := getFriendCode()
+	IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 	if(star = "Crown" || star = "Immersive")
 		RemoveFriends()
 	logMessage := star . " found by " . username . " (" . friendCode . ") in instance: " . scriptName . " (" . packs . " packs)\nFile name: " . accountFile . "\nBacking up to the Accounts\\SpecificCards folder and continuing..."
@@ -1231,7 +1272,7 @@ FindBorders(prefix) {
 }
 
 FindGodPack() {
-	global winTitle, discordUserId, Delay, username, packs, minStars
+	global winTitle, discordUserId, Delay, username, packs, minStars, scriptName, DeadCheck
 	gpFound := false
 	invalidGP := false
 	searchVariation := 5
@@ -1283,6 +1324,7 @@ FindGodPack() {
 				gpFound := true
 				GodPackFound("Invalid")
 				RemoveFriends()
+				IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 				break
 			}
 			else {
@@ -1296,7 +1338,9 @@ FindGodPack() {
 }
 
 GodPackFound(validity) {
+	global scriptName, DeadCheck
 	if(validity = "Valid") {
+		IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
 		Praise := ["Congrats!", "Congratulations!", "GG!", "Whoa!", "Praise Helix! ༼ つ ◕_◕ ༽つ", "Way to go!", "You did it!", "Awesome!", "Nice!", "Cool!", "You deserve it!", "Keep going!", "This one has to be live!", "No duds, no duds, no duds!", "Fantastic!", "Bravo!", "Excellent work!", "Impressive!", "You're amazing!", "Well done!", "You're crushing it!", "Keep up the great work!", "You're unstoppable!", "Exceptional!", "You nailed it!", "Hats off to you!", "Sweet!", "Kudos!", "Phenomenal!", "Boom! Nailed it!", "Marvelous!", "Outstanding!", "Legendary!", "Youre a rock star!", "Unbelievable!", "Keep shining!", "Way to crush it!", "You're on fire!", "Killing it!", "Top-notch!", "Superb!", "Epic!", "Cheers to you!", "Thats the spirit!", "Magnificent!", "Youre a natural!", "Gold star for you!", "You crushed it!", "Incredible!", "Shazam!", "You're a genius!", "Top-tier effort!", "This is your moment!", "Powerful stuff!", "Wicked awesome!", "Props to you!", "Big win!", "Yesss!", "Champion vibes!", "Spectacular!"]
 		invalid := ""
 	} else {
@@ -1314,6 +1358,10 @@ GodPackFound(validity) {
 	LogToFile(logMessage, godPackLog)
 	CreateStatusMessage(logMessage)
 	friendCode := getFriendCode()
+	IniWrite, 0, %A_ScriptDir%\%scriptName%.ini, UserSettings, DeadCheck
+	if(validity = "Valid") {
+		changeProfile()
+	}
 	logMessage := Interjection . "\n" . username . " (" . friendCode . ")\n[" . starCount . "/5][" . packs . "P] " . invalid . " God pack found in instance: " . scriptName . "\nFile name: " . accountFile . "\nBacking up to the Accounts\\GodPacks folder and continuing..."
 	LogToFile(logMessage, godPackLog)
 	;Run, http://google.com, , Hide ;Remove the ; at the start of the line and replace your url if you want to trigger a link when finding a god pack.
@@ -2036,6 +2084,7 @@ Delay(n) {
 }
 
 DoTutorial() {
+	global heartBeatName
 	FindImageAndClick(105, 396, 121, 406, , "Country", 143, 370) ;select month and year and click
 
 	Delay(1)
@@ -2203,15 +2252,22 @@ DoTutorial() {
 			adbClick(41, 296)
 		}
 	FindImageAndClick(190, 241, 225, 270, , "Name", 189, 438) ;wait for name input screen
+	FindImageAndClick(80, 448, 217, 480, , "OK7", 143, 179, 500) ; change profile image
+	Delay(3)
+	adbClick(143,244)
+	Delay(3)
+	adbClick(230,338)
+	Delay(3)
+	adbClick(143,466)
+	Delay(1)
 
 	FindImageAndClick(0, 476, 40, 502, , "OK", 139, 257) ;wait for name input screen
 
 	failSafe := A_TickCount
 	failSafeTime := 0
 	Loop {
-		name := ReadFile("usernames")
-		Random, randomIndex, 1, name.MaxIndex()
-		username := name[randomIndex]
+		Random, randomNum, 1, 3000  ; 生成 1 到 3000 的随机数
+		username := heartBeatName . randomNum  ; 拼接成 dino数字
 		username := SubStr(username, 1, 14)  ;max character limit
 		adbInput(username)
 		if(FindImageAndClick(121, 490, 161, 520, , "Return", 185, 372, , 10)) ;click through until return button on open pack
